+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/service/DriverSupportMaintainServiceImpl.java
@@ -1,20 +1,27 @@
 package com.dwarfeng.statistics.impl.service;
 
+import com.dwarfeng.statistics.impl.handler.DriverSupporter;
 import com.dwarfeng.statistics.stack.bean.entity.DriverSupport;
 import com.dwarfeng.statistics.stack.service.DriverSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
 import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
+import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import com.dwarfeng.subgrade.stack.log.LogLevel;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Component
 public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainService {
@@ -23,14 +30,22 @@ public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainSe
     private final DaoOnlyEntireLookupService<DriverSupport> entireLookupService;
     private final DaoOnlyPresetLookupService<DriverSupport> presetLookupService;
 
+    private final List<DriverSupporter> driverSupporters;
+
+    private final ServiceExceptionMapper sem;
+
     public DriverSupportMaintainServiceImpl(
             GeneralBatchCrudService<StringIdKey, DriverSupport> crudService,
             DaoOnlyEntireLookupService<DriverSupport> entireLookupService,
-            DaoOnlyPresetLookupService<DriverSupport> presetLookupService
+            DaoOnlyPresetLookupService<DriverSupport> presetLookupService,
+            List<DriverSupporter> driverSupporters,
+            ServiceExceptionMapper sem
     ) {
         this.crudService = crudService;
         this.entireLookupService = entireLookupService;
         this.presetLookupService = presetLookupService;
+        this.driverSupporters = Optional.ofNullable(driverSupporters).orElse(Collections.emptyList());
+        this.sem = sem;
     }
 
     @Override
@@ -288,5 +303,24 @@ public class DriverSupportMaintainServiceImpl implements DriverSupportMaintainSe
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public int lookupCount(String preset, Object[] objs) throws ServiceException {
         return presetLookupService.lookupCount(preset, objs);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    public void reset() throws ServiceException {
+        try {
+            List<StringIdKey> driverKeys = entireLookupService.lookupAsList().stream()
+                    .map(DriverSupport::getKey).collect(Collectors.toList());
+            crudService.batchDelete(driverKeys);
+            List<DriverSupport> driverSupports = driverSupporters.stream().map(supporter -> new DriverSupport(
+                    new StringIdKey(supporter.provideType()),
+                    supporter.provideLabel(),
+                    supporter.provideDescription(),
+                    supporter.provideExampleParam()
+            )).collect(Collectors.toList());
+            crudService.batchInsert(driverSupports);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logParse("重置映射器支持时发生异常", LogLevel.WARN, e, sem);
+        }
     }
 }