+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/service/FilterSupportMaintainServiceImpl.java
@@ -1,20 +1,27 @@
 package com.dwarfeng.statistics.impl.service;
 
+import com.dwarfeng.statistics.impl.handler.FilterSupporter;
 import com.dwarfeng.statistics.stack.bean.entity.FilterSupport;
 import com.dwarfeng.statistics.stack.service.FilterSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
 import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
+import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import com.dwarfeng.subgrade.stack.log.LogLevel;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Component
 public class FilterSupportMaintainServiceImpl implements FilterSupportMaintainService {
@@ -23,14 +30,22 @@ public class FilterSupportMaintainServiceImpl implements FilterSupportMaintainSe
     private final DaoOnlyEntireLookupService<FilterSupport> entireLookupService;
     private final DaoOnlyPresetLookupService<FilterSupport> presetLookupService;
 
+    private final List<FilterSupporter> filterSupporters;
+
+    private final ServiceExceptionMapper sem;
+
     public FilterSupportMaintainServiceImpl(
             GeneralBatchCrudService<StringIdKey, FilterSupport> crudService,
             DaoOnlyEntireLookupService<FilterSupport> entireLookupService,
-            DaoOnlyPresetLookupService<FilterSupport> presetLookupService
+            DaoOnlyPresetLookupService<FilterSupport> presetLookupService,
+            List<FilterSupporter> filterSupporters,
+            ServiceExceptionMapper sem
     ) {
         this.crudService = crudService;
         this.entireLookupService = entireLookupService;
         this.presetLookupService = presetLookupService;
+        this.filterSupporters = Optional.ofNullable(filterSupporters).orElse(Collections.emptyList());
+        this.sem = sem;
     }
 
     @Override
@@ -288,5 +303,24 @@ public class FilterSupportMaintainServiceImpl implements FilterSupportMaintainSe
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public int lookupCount(String preset, Object[] objs) throws ServiceException {
         return presetLookupService.lookupCount(preset, objs);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    public void reset() throws ServiceException {
+        try {
+            List<StringIdKey> filterKeys = entireLookupService.lookupAsList().stream()
+                    .map(FilterSupport::getKey).collect(Collectors.toList());
+            crudService.batchDelete(filterKeys);
+            List<FilterSupport> filterSupports = filterSupporters.stream().map(supporter -> new FilterSupport(
+                    new StringIdKey(supporter.provideType()),
+                    supporter.provideLabel(),
+                    supporter.provideDescription(),
+                    supporter.provideExampleParam()
+            )).collect(Collectors.toList());
+            crudService.batchInsert(filterSupports);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logParse("重置映射器支持时发生异常", LogLevel.WARN, e, sem);
+        }
     }
 }