+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/service/ProviderSupportMaintainServiceImpl.java
@@ -1,20 +1,27 @@
 package com.dwarfeng.statistics.impl.service;
 
+import com.dwarfeng.statistics.impl.handler.ProviderSupporter;
 import com.dwarfeng.statistics.stack.bean.entity.ProviderSupport;
 import com.dwarfeng.statistics.stack.service.ProviderSupportMaintainService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
 import com.dwarfeng.subgrade.impl.service.DaoOnlyPresetLookupService;
 import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
+import com.dwarfeng.subgrade.sdk.exception.ServiceExceptionHelper;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
 import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import com.dwarfeng.subgrade.stack.exception.ServiceExceptionMapper;
+import com.dwarfeng.subgrade.stack.log.LogLevel;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Collections;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Component
 public class ProviderSupportMaintainServiceImpl implements ProviderSupportMaintainService {
@@ -23,14 +30,22 @@ public class ProviderSupportMaintainServiceImpl implements ProviderSupportMainta
     private final DaoOnlyEntireLookupService<ProviderSupport> entireLookupService;
     private final DaoOnlyPresetLookupService<ProviderSupport> presetLookupService;
 
+    private final List<ProviderSupporter> providerSupporters;
+
+    private final ServiceExceptionMapper sem;
+
     public ProviderSupportMaintainServiceImpl(
             GeneralBatchCrudService<StringIdKey, ProviderSupport> crudService,
             DaoOnlyEntireLookupService<ProviderSupport> entireLookupService,
-            DaoOnlyPresetLookupService<ProviderSupport> presetLookupService
+            DaoOnlyPresetLookupService<ProviderSupport> presetLookupService,
+            List<ProviderSupporter> providerSupporters,
+            ServiceExceptionMapper sem
     ) {
         this.crudService = crudService;
         this.entireLookupService = entireLookupService;
         this.presetLookupService = presetLookupService;
+        this.providerSupporters = Optional.ofNullable(providerSupporters).orElse(Collections.emptyList());
+        this.sem = sem;
     }
 
     @Override
@@ -288,5 +303,24 @@ public class ProviderSupportMaintainServiceImpl implements ProviderSupportMainta
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
     public int lookupCount(String preset, Object[] objs) throws ServiceException {
         return presetLookupService.lookupCount(preset, objs);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    public void reset() throws ServiceException {
+        try {
+            List<StringIdKey> providerKeys = entireLookupService.lookupAsList().stream()
+                    .map(ProviderSupport::getKey).collect(Collectors.toList());
+            crudService.batchDelete(providerKeys);
+            List<ProviderSupport> providerSupports = providerSupporters.stream().map(supporter -> new ProviderSupport(
+                    new StringIdKey(supporter.provideType()),
+                    supporter.provideLabel(),
+                    supporter.provideDescription(),
+                    supporter.provideExampleParam()
+            )).collect(Collectors.toList());
+            crudService.batchInsert(providerSupports);
+        } catch (Exception e) {
+            throw ServiceExceptionHelper.logParse("重置映射器支持时发生异常", LogLevel.WARN, e, sem);
+        }
     }
 }