+++ statistics-stack/src/main/java/com/dwarfeng/statistics/stack/service/MapperSupportMaintainService.java
@@ -2,6 +2,7 @@ package com.dwarfeng.statistics.stack.service;
 
 import com.dwarfeng.statistics.stack.bean.entity.MapperSupport;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
 import com.dwarfeng.subgrade.stack.service.BatchCrudService;
 import com.dwarfeng.subgrade.stack.service.EntireLookupService;
 import com.dwarfeng.subgrade.stack.service.PresetLookupService;
@@ -17,4 +18,11 @@ public interface MapperSupportMaintainService extends BatchCrudService<StringIdK
 
     String ID_LIKE = "id_like";
     String LABEL_LIKE = "label_like";
+
+    /**
+     * 重置映射器支持。
+     *
+     * @throws ServiceException 服务异常。
+     */
+    void reset() throws ServiceException;
 }